+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [Unreleased]
+- fix bug in `simplebot.utils.get_default_account()` (#72)
+
 ## [1.1.0]
 
 - Improved pytestplugin to allow simulating incoming messages with encryption errors (#68)
@@ -27,7 +30,7 @@
 - initial release
 
 
-[unreleased]: https://github.com/simplebot-org/simplebot/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/simplebot-org/simplebot/compare/v1.1.0...HEAD
 
 [1.1.0]: https://github.com/simplebot-org/simplebot/compare/v1.0.1...v1.1.0